Refactor: extract shared process utilities and add non-blocking pipe write
- Create process.py with shared CommandError, CommandTimeoutError,
  wait_for_process(), and kill_process() functions
- Update sound.py and tts.py to use shared process module
- Add _write_to_pipe_nonblocking() in tts.py to handle large input
  without blocking indefinitely (uses select with timeout)
- Eliminates code duplication between sound.py and tts.py
