+++ src/audio_notify_server/sound.py
@@ -2,78 +2,22 @@
 
 from __future__ import annotations
 
-import contextlib
 import os
 import shutil
-import signal
 import sys
-import time
 from pathlib import Path
+
+from audio_notify_server.process import (
+    CommandError,
+    CommandTimeoutError,
+    wait_for_process,
+)
 
 # Timeout for audio playback operations (in seconds)
 AUDIO_PLAYBACK_TIMEOUT = 10
 
 # Trusted audio player executables - hardcoded list for security
 TRUSTED_AUDIO_PLAYERS = frozenset({"paplay", "pw-play", "aplay", "ffplay", "mpv"})
-
-
-class CommandError(Exception):
-    """Command execution error."""
-
-
-class CommandTimeoutError(Exception):
-    """Command timeout error."""
-
-
-def _wait_for_process(pid: int, timeout: float) -> int:
-    """Wait for process to complete with timeout.
-
-    Args:
-        pid: Process ID to wait for.
-        timeout: Timeout in seconds.
-
-    Returns:
-        Exit code of the process.
-
-    Raises:
-        CommandError: If the process fails.
-        CommandTimeoutError: If the process times out.
-
-    """
-    start_time = time.time()
-    while True:
-        try:
-            wpid, status = os.waitpid(pid, os.WNOHANG)
-        except ChildProcessError:
-            # Process doesn't exist (already reaped)
-            return 0
-        if wpid == pid:
-            if os.WIFEXITED(status):
-                return os.WEXITSTATUS(status)
-            msg = "Command terminated abnormally"
-            raise CommandError(msg)
-
-        if time.time() - start_time > timeout:
-            _kill_process(pid)
-            msg = f"Command timed out after {timeout} seconds"
-            raise CommandTimeoutError(msg)
-
-        time.sleep(0.01)
-
-
-def _kill_process(pid: int) -> None:
-    """Kill a process gracefully then forcefully.
-
-    Args:
-        pid: Process ID to kill.
-
-    """
-    with contextlib.suppress(ProcessLookupError):
-        os.kill(pid, signal.SIGTERM)
-        time.sleep(0.1)
-        os.kill(pid, signal.SIGKILL)
-    with contextlib.suppress(ChildProcessError):
-        os.waitpid(pid, 0)
 
 
 def _safe_run_audio_command(
@@ -134,7 +78,7 @@ def _safe_run_audio_command(
         os.close(devnull_fd)
 
     # Wait for process with timeout
-    exit_code = _wait_for_process(pid, timeout)
+    exit_code = wait_for_process(pid, timeout)
     if exit_code != 0:
         msg = f"Command failed with exit code {exit_code}"
         raise CommandError(msg)
